Use zustand for schema state
