+++ client/package.json
@@ -46,7 +46,8 @@
     "unistore": "^3.5.2",
     "use-debounce": "^5.0.0",
     "uuid": "^8.3.0",
-    "whatwg-fetch": "^3.4.1"
+    "whatwg-fetch": "^3.4.1",
+    "zustand": "^3.1.3"
   },
   "scripts": {
     "analyze": "source-map-explorer \"build/static/js/*.js\"",
