+++ client/src/queryEditor/toolbar/ToolbarToggleSchemaButton.tsx
@@ -1,11 +1,9 @@
 import DatabaseIcon from 'mdi-react/DatabaseIcon';
 import React from 'react';
 import IconButton from '../../common/IconButton';
-import useSchemaState from '../../stores/use-schema-state';
+import { toggleSchema } from '../../stores/schema-store';
 
 function ToolbarToggleSchemaButton() {
-  const { toggleSchema } = useSchemaState();
-
   return (
     <IconButton tooltip="Toggle schema" onClick={toggleSchema}>
       <DatabaseIcon />
